requirements.txt
asgiref==3.4.1
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
city==0.1
click==8.0.3
djando==0.2
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kwargs==1.0.1
legacy==0.1.6
lxml==4.6.4
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
psycopg2-binary==2.9.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
street==0.1.2
urllib3==1.26.7
Werkzeug==2.0.2
